Add routeData for videos. Update videos display.
